// bo-front/src/app/profiles/pages/ambassador/pages/tree/commons/interfaces/sponsorsList.ts
export class SponsorsListResponseDTO{
    idUser: number;
    fullName: string;
    userName: string;
    nroDocument: string;
    idState: number;
    stateName: string;
    colorRGB: string;
    createDate: string | null;
    sponsorLevel: number;
    sponsorName: string;
    rangeId: number;
    rangeName: string;
    directPartnersCount: number;
    directSponsorshipScore: number
}